% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{create_authentication_key}
\alias{create_authentication_key}
\title{Create Authentication Key to be used with all API calls.}
\usage{
create_authentication_key(username,
  key = rstudioapi::askForSecret("Key"))
}
\arguments{
\item{username}{The email registered with Azure DevOps}

\item{key}{Password or PAT Code}
}
\value{
Authentication Key
}
\description{
Create Authentication Key to be used with all API calls.
}
\examples{
\dontrun{
create_authentication_key("adam.robinson@education.gov.uk")
}
}
\keyword{Authenticate}
