% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_status_report}
\alias{generate_status_report}
\title{Creates an html report summarising all projects and repos for DevOps Site.}
\usage{
generate_status_report(domain, auth,
  output_file = "azuredevops_status_report.html", output_dir = ".")
}
\arguments{
\item{domain}{DevOps Site Name}

\item{auth}{Authentication Key}

\item{output_file}{Name of output file}

\item{output_dir}{Name of output directory}
}
\value{
HTML Report
}
\description{
Creates an html report summarising all projects and repos for DevOps Site.
}
\examples{
\dontrun{
generate_status_report("teachers-analysis-unit", auth)
}
}
\keyword{Branches}
