% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builds.R
\name{get_build_definition}
\alias{get_build_definition}
\title{Get build defintion.}
\usage{
get_build_definition(definitionid, project, domain, auth)
}
\arguments{
\item{definitionid}{Build Definition ID}

\item{project}{DevOps Project Name}

\item{domain}{DevOps Site Name}

\item{auth}{Authentication Ket}
}
\value{
JSON Definition
}
\description{
Get build defintion.
}
\examples{
\dontrun{
get_build_definition(1, "sub-national-supply", "teachers-analysis-unit", auth)
}
}
\keyword{Build}
\keyword{Builds}
\keyword{Defs,}
