% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{list_branches}
\alias{list_branches}
\title{List all branches within a Repo.}
\usage{
list_branches(repo, project, domain, auth)
}
\arguments{
\item{repo}{Repository name}

\item{project}{DevOps Project Name}

\item{domain}{DevOps Site Name}

\item{auth}{Authentication Key}
}
\value{
Dataframe
}
\description{
List all branches within a Repo.
}
\examples{
\dontrun{
list_branches("supply-index", "sub-national-supply", "teachers-analysis-unit", auth)
}
}
\keyword{Branches}
