% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{list_projects}
\alias{list_projects}
\title{List all projects.}
\usage{
list_projects(domain, auth)
}
\arguments{
\item{domain}{The name of the DevOps Site (short name)}

\item{auth}{Authentication Variable}
}
\value{
Dataframe
}
\description{
List all projects.
}
\examples{
\dontrun{
list_projects("teachers-analysis-unit", auth)
}
}
\keyword{Projects}
