% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{list_repos}
\alias{list_repos}
\title{List all repos within a project}
\usage{
list_repos(project, domain, auth)
}
\arguments{
\item{project}{DevOps Project Name}

\item{domain}{DevOps Site Name}

\item{auth}{Authentication Key}
}
\value{
Dataframe
}
\description{
List all repos within a project
}
\examples{
\dontrun{
list_repos("sub-national-supply", "teachers-analysis-unit", auth)
}
}
\keyword{Projects}
\keyword{Repos,}
