% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builds.R
\name{queue_build}
\alias{queue_build}
\title{Queue a build based on existing definition.}
\usage{
queue_build(definitionid, project, domain, auth)
}
\arguments{
\item{definitionid}{Build definition ID}

\item{project}{Project Name}

\item{domain}{DevOps Site Name}

\item{auth}{Authentication Key}
}
\value{
Dataframe
}
\description{
Queue a build based on existing definition.
}
\examples{
\dontrun{
queue_build(1, "sub-national-supply", "teachers-analysis-unit", auth)
}
}
\keyword{Builds}
